ECS Finished and tested. System can be fleshed out a bit more eventually, can also optimize Entitys using object pools.
