+++ bobo_engine/src/Bobo/ECS/BaseComponents/Transform.h
@@ -1,9 +1,9 @@
 #pragma once
 
 #include <glm/glm.hpp>
-#include "Bobo/ECS2/Component.h"
+#include "Bobo/ECS/Component.h"
 
-namespace Bobo2
+namespace Bobo
 {
 	struct Transform : public Component
 	{
