--- bobo_engine/src/Bobo/ECS/Scene.cpp
@@ -1,2 +1,0 @@
-#include "bobopch.h"
-#include "Scene.h"
