--- bobo_engine/src/Bobo/ECS2/Component.h
@@ -1,10 +1,0 @@
-#pragma once
-
-namespace Bobo2
-{
-	class Component
-	{
-	public:
-		virtual ~Component() = default;
-	};
-}
